+++ src/components/feature/Auth/LoginForm/SendEmailModal.tsx
@@ -2,8 +2,6 @@
 
 import { useSendResetPasswordMutation } from "@/api/user/user.query";
 import PasswordResetModal from "@/components/common/Modal/PasswordResetModal";
-import { useModalStore } from "@/stores/modalStore";
-import { useToastStore } from "@/stores/toastStore";
 import React, { useState } from "react";
 
 export default function SendEmailModal() {
@@ -11,8 +9,6 @@ export default function SendEmailModal() {
   const sendEmailMutation = useSendResetPasswordMutation();
   const redirectUrl =
     typeof window !== "undefined" ? window.location.origin : "";
-  const { closeModal } = useModalStore();
-  const { showToast } = useToastStore();
 
   const handleEmailChange = (e: React.ChangeEvent<HTMLInputElement>) => {
     setEmail(e.target.value);
@@ -24,12 +20,7 @@ export default function SendEmailModal() {
       { email, redirectUrl: redirectUrl },
       {
         onSuccess: () => {
-          showToast("비밀번호 재설정 링크가 전송되었습니다.", "success");
-          closeModal();
           setEmail("");
-        },
-        onError: () => {
-          showToast("이메일을 확인해주세요.", "error");
         },
       },
     );
